Randomly schedule metadata updates
The backgroundMetadataUpdater task uses a simple time.Ticker that
triggers a periodic metadata refresh for a kafka client (once every 10
minutes by default). The metadata refresh itself communicates with the
kafka cluster, and we’re unable to produce/consume messages for a short
time as a result. This is usually fine though since the refresh is
infrequent, and is not a major delay.

However, when running many instances of a consumer/producer (like we do
at VividCortex) across a datacenter, metadata refreshes can trigger
minor storms of traffic when synchronized across all of the kafka
clients simultaneously. This can be mitigated rather easily by
introducing some randomization into the metadata refresh task.

This change adds a random sleep, up to half of the
Metadata.RefreshFrequency before every metadata update. Doing so should
mitigate the possibility for storms as a result of synchronized
updates, as it is far more likely for only 1 or two clients to update
their metadata in a datacenter concurrently (as opposed to when the
task is done predictably across all clients that are simultaneously
deployed).

The randomization added here is disabled by default, and can be enabled
by setting the Client’s Config.Metadata.RandomRefresh flag.